review updater.sh
added a condition to choose if clean the trash or not

+++ updater.sh
@@ -99,6 +99,15 @@ else
     printf "${RED}System not supported${NORMAL}"
 fi
 
-sudo rm -rf ~/.local/share/Trash/*
+printf "${RED}Showing files in .local/share/Trash/${NORMAL}\n"
+ls ~/.local/share/Trash/
+printf "Do you wanna clean trash?"
+read -p "[y/n]: " answer 
+if [[ "$answer" == "y" ]]; then
+    sudo rm -rf ~/.local/share/Trash/*
+    printProgress trash: cleaned
+else
+    printProgress trash: not cleaned
+fi
 
 printf "\n"
